Changed TextMessages to use longs for date and switched from absolute to relative date cursor info

+++ src/com/whitaker/textalyzer/TextMessage.java
@@ -8,7 +8,7 @@ public class TextMessage
 	
 	public long timeCreated; 
 
-	public TextMessage(Directions direction, String body, int timeCreated) 
+	public TextMessage(Directions direction, String body, long timeCreated) 
 	{
 		this.direction = direction;
 		this.body = body;
